Major refactoring and bump react-scripts to 5.0.1
