--- frontend/package.json
@@ -1,1 +1,0 @@
-/home/santeri/Documents/anecdotes-cicd/package.json
