+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import store from './store';
+import store from './src/store';
 import { Provider } from 'react-redux';
-import App from './App';
+import App from './src/App';
 
 ReactDOM.render(
   <Provider store={store}>
